Corregido organizacion de paquete spamfilter
